--- src/app/app.routes.js
@@ -1,9 +1,0 @@
-import { appComponent } from './app.component.js';
-import { signinComponent } from './auth.signincomponent.js';
-import { signupComponent } from './auth.signupcomponent.js';
-
-export const appRoutes = [
-  { path: '/', component: appComponent },
-  { path: '/signin', component: signupComponent },
-  { path: '/signup', component: signinComponent },
-];
