+++ src/framework/core/regex.js
